+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- add `UIApplication.shared.legacyKeyWindow` to get `keyWindow` for scene-based apps.
+
 ## [1.1.3] - 2020-02-23
 
 ### Changed